+++ test/esModule-option.test.js
@@ -250,4 +250,56 @@ describe('"esModule" option', () => {
       expect(getErrors(stats)).toMatchSnapshot("errors");
     });
   });
+
+  const injectTypesForNamedExport = [
+    "styleTag",
+    "singletonStyleTag",
+    "lazyStyleTag",
+    "lazySingletonStyleTag",
+  ];
+
+  injectTypesForNamedExport.forEach((injectType) => {
+    it(`should work with the "${injectType}" inject type`, async () => {
+      expect.assertions(3);
+
+      const compiler = getCompiler(
+        "./named-export.js",
+        {},
+        {
+          module: {
+            rules: [
+              {
+                test: /\.css$/i,
+                use: [
+                  {
+                    loader: path.resolve(__dirname, "../src/cjs.js"),
+                    options: {
+                      injectType,
+                      esModule: true,
+                    },
+                  },
+                  {
+                    loader: "css-loader",
+                    options: {
+                      modules: {
+                        namedExport: true,
+                      },
+                    },
+                  },
+                ],
+              },
+            ],
+          },
+        }
+      );
+      const stats = await compile(compiler);
+
+      runInJsDom("main.bundle.js", compiler, stats, (dom) => {
+        expect(dom.serialize()).toMatchSnapshot("DOM");
+      });
+
+      expect(getWarnings(stats)).toMatchSnapshot("warnings");
+      expect(getErrors(stats)).toMatchSnapshot("errors");
+    });
+  });
 });
